+++ src/main/java/br/com/maurigvs/surveyapi/repository/QuestionRepository.java
@@ -1,10 +1,9 @@
 package br.com.maurigvs.surveyapi.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import br.com.maurigvs.surveyapi.model.Question;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 }
